requirements.txt
fastapi
uvicorn
python-multipart
pymongo
certifi
python-dotenv
pytesseract
opencv-python
pandas
pillow
numpy
joblib
scikit-learn
spacy
bcrypt
